WIP: Reorganize project.
