+++ openbb_agents/agents/v1/__init__.py
@@ -1,0 +1,1 @@
+from .agent import openbb_agent
