+++ openbb_agents/agents/v2/__init__.py
@@ -1,0 +1,1 @@
+from .agent import openbb_agent
