+++ pyproject.toml
@@ -18,7 +18,6 @@ langchainhub = "^0.1.14"
 python-json-logger = "^2.0.7"
 fastapi = "^0.104.1"
 uvicorn = "^0.24.0"
-argparse = "^1.4.0"
 matplotlib = "^3.8.2"
 openbb-charting = "^1.0.0"
 
@@ -29,3 +28,6 @@ pre-commit = "^3.5.0"
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.isort]
+profile="black"
